学习eslint+babel+es特性以及配置项目文件格式=-=

+++ server/models/users.js
@@ -7,7 +7,8 @@ export async function createUserTable() {
       password    VARCHAR(16) NOT NULL,
       telephone   VARCHAR(20) UNIQUE NOT NULL,
       email       VARCHAR(255) UNIQUE NOT NULL,
-      is_manager  BOOL DEFAULT false NOT NULL
+      is_manager  BOOL DEFAULT false NOT NULL,
+      test        INT DEFAULT 0
       )`, undefined, "Table users created...");
 }
 
@@ -16,24 +17,24 @@ export async function dropUserTable() {
 }
 
 export async function createUser(user) {
-  return await execAsync("INSERT INTO users " +
-    "(user_id, username, password, telephone, email) VALUES (?, ?, ?, ?, ?)",
+  return await execAsync(`INSERT INTO users 
+    (user_id, username, password, telephone, email) VALUES (?, ?, ?, ?, ?)`,
     [user.user_id, user.username, user.password, user.telephone, user.email],
     "create user " + JSON.stringify(user));
 }
 
 // for login by email
 export async function getUserByEmail(email, password) {
-  return await execAsync("SELECT user_id, username, email, telephone, is_manager FROM users " +
-    "WHERE email = ?",
+  return await execAsync(`SELECT user_id, username, email, telephone, is_manager FROM users 
+    WHERE email = ?`,
     [email],
     `select user by email ${email}`);
 }
 
 // for login by phone
 export async function getUserByPhone(telephone, password) {
-  return await execAsync("SELECT user_id, username, email, telephone, is_manager FROM users " +
-    "WHERE telephone = ?",
+  return await execAsync(`SELECT user_id, username, email, telephone, is_manager FROM users 
+    WHERE telephone = ?`,
     [telephone],
     `select user by telephone ${telephone}`);
 }
@@ -45,29 +46,25 @@ export async function getUsernameByUserID(user_id) {
 }
 
 export async function changePassword(new_password, user_id, password) {
-  return await execAsync("UPDATE users SET password = ? " +
-  "WHERE user_id = ? AND password = ?",
-  [new_password, user_id, password],
-  `update user [user_id: ${user_id}, password: ${password}]`);
+  return await execAsync("UPDATE users SET password = ? WHERE user_id = ? AND password = ?",
+    [new_password, user_id, password],
+    `update user [user_id: ${user_id}, password: ${password}]`);
 }
 
 export async function changeEmail(new_email, user_id, email) {
-  return await execAsync("UPDATE users SET email = ? " +
-  "WHERE user_id = ? AND email = ?",
-  [new_email, user_id, email],
-  `update user [user_id: ${user_id}, email: ${email}]`);
+  return await execAsync("UPDATE users SET email = ? WHERE user_id = ? AND email = ?",
+    [new_email, user_id, email],
+    `update user [user_id: ${user_id}, email: ${email}]`);
 }
 
 export async function changeTelephone(new_telephone, user_id, telephone) {
-  return await execAsync("UPDATE users SET telephone = ? " +
-  "WHERE user_id = ? AND telephone = ?",
-  [new_telephone, user_id, telephone],
-  `update user [user_id: ${user_id}, telephone: ${telephone}]`);
+  return await execAsync("UPDATE users SET telephone = ? WHERE user_id = ? AND telephone = ?",
+    [new_telephone, user_id, telephone],
+    `update user [user_id: ${user_id}, telephone: ${telephone}]`);
 }
 
 export async function deleteUser(user_id, password) {
-  return await execAsync("DELETE FROM users " +
-    "WHERE user_id = ? AND password = ?",
+  return await execAsync("DELETE FROM users WHERE user_id = ? AND password = ?",
     [user_id, password],
     `delete user [user_id: ${user_id}]`);
 }
